Fixed unit test not seeing file in small data.

+++ h2o-algos/src/test/java/hex/glm/GLMBasicTest.java
@@ -89,8 +89,9 @@ public class GLMBasicTest extends TestUtil {
   @BeforeClass
   public static void setup() {
     stall_till_cloudsize(1);
-    File f = new File("smalldata/glm_test/prostate_cat_replaced.csv");
+    File f = find_test_file("smalldata/glm_test/prostate_cat_replaced.csv");
     assert f.exists();
+
     NFSFileVec nfs = NFSFileVec.make(f);
     Key outputKey = Key.make("prostate_test.hex");
     _prostate = ParseDataset.parse(outputKey, nfs._key);
@@ -98,6 +99,7 @@ public class GLMBasicTest extends TestUtil {
 
   @AfterClass
   public static void cleanUp() {
-    _prostate.delete();
+    if(_prostate != null)
+      _prostate.delete();
   }
 }
